requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
streamlit~=1.41.1
pillow~=11.1.0
streamlit-drawable-canvas~=1.17.0